Added auto mode
If you toggle the thingy, a single search result will automatically be
added to your deck. This is also on by default for the version selector,
and the version selector now focuses its filter field, so you can
alt-click then type "s=ust<ENTER>" to select the Unstable version of the
basic lands.

And you should.

Signed-off-by: Emi <dealuminis@gmail.com>