Start isolating access to CardView model
Concurrent access to this is a huge problem... I'm preparing to turn
CardView into a more asynchronous component able to accept model updates
and repaint the view on the fly, but to do that I need to make sure the
model is never updated from more than one thread at a time -- which
means very carefully controlling access to it. The model value is going
to become package-private and I'm going to remove the accessors, so this
is the first step -- since groupings aren't in the same package, they
need to accept the model as a parameter (even though that's mostly
unused).

Signed-off-by: Emi <dealuminis@gmail.com>