Remove direct access to CardView.model and .filteredModel
Open access to this was causing concurrent modification issues when I
made CardView rendering asynchronous. All changes should be made
synchronized on the unfiltered list so that listeners for filtered lists
never trample each other; JavaFX does not handle comod well at all. The
solution to this was to have a single method capable of mutating the
model, accessible through cardPane.changeModel().

Signed-off-by: Emi <dealuminis@gmail.com>