Fix CardView render concurrent cancelation system.
Dunno why the original version was so complicated. Before, though it
would attempt to cancel a prescheduled render job, that only worked if
the common fork/join pool hadn't already started in on the render code.
The new method is simple and elegant: keep a volatile long to track the
most recently scheduled render number and bail early if it ever exceeds
the number handed to the render code at scheduling time.

Multiple calls to scheduleRender now increment the render generation,
causing any previously scheduled render jobs to bail out at a few
different break points, such as before building the render map or before
scheduling the task on the application thread, or at the time the
application thread goes to do the drawing.

This is nice for a lot of reasons. For the most part, only the *last*
scheduled render is likely to complete, and should *never* complete
before a previously scheduled render (so the view never flickers with
old data or the like).

Of course, in theory, it's possible for the number of renders in the
lifetime of the application to exceed the largest value of a long. I did
try to catch this case and, if an opportunity arises, reset the
generation ID to 0. Hopefully that never comes up though. It's like 4.5
quintillion renders before I have to worry about that.

Weirdly, Idea seems to think the repeated if statements are a code error
or bad code or something; I guess it doesn't check that the conditional
involves a volatile variable which might be changed by other threads?

Anyway this should help a lot with async rendering, since only the most
recent scheduled render will eventually complete.

Signed-off-by: Emi <dealuminis@gmail.com>