Add image path preference
You can now specify the images path in the deckbuilder preferences. The
change does require you to restart the deckbuilder though, since the
images object is still locked to the paths at initialization.

I may change this later, but it's important to realize that the old card
images aren't moved...

Signed-off-by: Emi <dealuminis@gmail.com>