Reduce flickering due to mouse movements
We were accidentally halfway rendering some cards from empty groups. We
also scheduled *tons* of renders for each card that had to load. There
are probably a few other tricks I used to reduce the calls to
scheduleRender and layout. But now the only flickering is due to the
model not being settled, which I can't really do much about. Unless I
maybe synchronize on the model...?

Signed-off-by: Emi <dealuminis@gmail.com>