Eliminate Import/Export features for serdes and allow multiple extensions
Kind of two sides of the same coin. Whether or not a serdes can import or export is now determined by whether or not its list of handled import/export extensions is empty. Asymmetric serdes like Images return an empty list for import extensions and however many export extensions it can handle.

Signed-off-by: Emi <dealuminis@gmail.com>