Significantly improve multi-screen windowing experience
This is a complicated chain of fixing host windows and positioning dialogs, et cetera. But it makes the multiple monitor experience way better. The startup dialogs will be moved to the same screen where the mouse pointer is on startup -- this isn't guaranteed to be perfect, but it should generally open on the right screen. All other dialogs should descend from whatever opens them, and new main windows are created near the window that opened them.

Signed-off-by: Emi <dealuminis@gmail.com>