Add MicroMarkdown markdown translator
I'm going to add an in-application changelog, and this will greatly simplify writing that changelog.

Honestly it's a really, really shitty implementation of not-quite-Markdown. But it's also roughly on the order of 1.5n (I think?) and can operate on a stream (though it needs a lot of lookahead). I'm happy with it, and that's enough for me.

Signed-off-by: Emi <dealuminis@gmail.com>