Centralize ad-hoc plugin loading procedures
This is one of those "huge backend, small changelog" updates. The short version is that an outdated plugin (looking at you, TTS export) won't cause the deckbuilder to crash on startup. The long version is that all services should now load providers through PluginUtils, optionally providing a linkage validator to exercise the API and make sure it hasn't changed since the plugin JAR was created. If it has, an error is printed to stderr.

In the future, I may require plugins to provide a specific Plugin service which can be used to validate API versions, update the plugin, et cetera, but all of that seems like ages away, since the only plugin at the moment is my own...

Signed-off-by: Emi <dealuminis@gmail.com>