Add changelog note for fixed download issues in libmtg-scryfall
Apparently Cloudflare tightened their security and began rejecting some requests from the deckbuilder. On the advice of the Scryfall support staff, all requests by that plugin now include a custom user-agent and an Accept header, which should avoid the Cloudflare blocks. Check the libmtg-scryfall repo for the relevant commit; this just adds a note to the deckbuilder changelog.

Signed-off-by: Emi <dealuminis@gmail.com>