Noted improvement to Scryfall plugin to avoid enum map key bricking
This was a big problem for a while; Scryfall would introduce a new enum constant and the library would choke on it. Adding Unrecognized helped for a little bit, but if they ever added two formats at once -- which has happened -- the plugin would stop working again. Now... they're just strings. It's slower than I'd like, but safer.

Signed-off-by: Emi <dealuminis@gmail.com>