Increase flexibility of UniqueList's preference system
Rather than "is this the preferred instance of the item", it's now "is this item more preferable to display than this other one". In theory it should be possible to use UniqueList further up the chain for showing versions separately, and this is meant to enable that -- but I'm not brave enough to tinker with the inner workings of that system right now.

Signed-off-by: Emi <dealuminis@gmail.com>