Add "automatic" cutboard option that removes cards to cutboard as long as it's open
This was a nice QoL feature recommended by nobodi ages ago that I'm finally getting around to. This is the default behavior now: double-clicking a card, or right-click removing it, will remove the card from the deck entirely, unless the cutboard is open -- in which case it will be transferred there instead.

Signed-off-by: Emi <dealuminis@gmail.com>