Json serdes: Improve resiliency against upstream set code/collector number changes
Scryfall has taken to changing around set IDs and collector numbers lately (especially with The List). This leads to decks sometimes becoming un-loadable, since the JSON importer barfs up an exception if it can't find the exact card match. I modified the printing deserializer to be kinder and gentler about that kind of thing; if it can't find that collector number in that set, it'll fall back to a printing of that card in that set; if it can't find the set, or a card by that name in the set, it'll find the preferred printing of a card with that name in the data source. (Failing that, though, it'll throw an exception again, but at least this time it's an IOException which gets caught by the MainWindow trying to open the deck, rather than caught by the uncaught-exception handler...)

Signed-off-by: Emi <dealuminis@gmail.com>