Add barebones debug console
Right now, it just shows the contents of stdout and stderr. But this can be really helpful in debugging certain weird problems, especially around loading card data; back in the day, I used to hand out a batch file to redirect stdout and stderr to files that people could upload. This shortens that loop a bit.

Signed-off-by: Emi <dealuminis@gmail.com>