Improve data source selection behavior
Auto-load is now determined by a boolean (defaults to true since for now there will only be one data source), and preferred data source is configurable in the preferences menu or on the data loading dialog. This all happened because I was working on Scryfall and made two separate versions of the Scryfall plugin -- one with JSON and one with MessagePack -- but I'm going to back that change out and use a newly-invented PluginPreferences system to control the choice of serializer. Please Look Forward To It.

Signed-off-by: Emi <dealuminis@gmail.com>