Plugin Preferences: Initial draft
Adds a mechanism by which external plugins can provide preferences to the user. Controls are automatically created (random order right now) based on an annotation applied to non-transient fields of the provider class. Getting GSON to serialize/deserialize this properly was kind of a nightmare, but I'm happier with the end result than I was at any point on the way there, so I'll take it.

Signed-off-by: Emi <dealuminis@gmail.com>