CardPane: Add combo box to select search provider much more easily.
I wonder how many people even know there's multiple search providers? This patch adds a combo box to the right end of the filter bar. It uses a stack pane to render the combo over the top of the search box, and although the box itself isn't mouse-transparent, it only displays if the bar is wide enough.

Signed-off-by: Emi <dealuminis@gmail.com>