Add simple logging (slogging, if you will) tool and use throughout
I wasn't sure I'd ever do this, but at this point, it's simply useful. There are a few usages that are dubious -- I'm not sure controller objects should be importing MainApplication just to use its root logger -- should they have their own...? -- but it's going to be nice to have for a while. I hope.

Signed-off-by: Emi <dealuminis@gmail.com>