ImageExporter: Separate out PNG and JPG, drop GIF, and implement CopyPaste
I've been wanting to be able to copy JPEGs for a while -- they tend to be much, much smaller files while still being readable, likely due to the mixture of illustrations and text -- but couldn't come up with the right approach. I'm still not sure this one is, but it's here now -- I'll make tweaks in the future if it feels necessary.

I also split PNG and JPG into separate image exporters, since they have very different clipboard-export mechanisms (PNGs still copy directly to your clipboard as bitmap data, though a full deck PNG tends to be >10MB, the size limit for a Discord file attachment).

Signed-off-by: Emi <dealuminis@gmail.com>