DeckImportExport: Introduce DataFormat interface and monotype all importers/exporters
Image import/export was the only exporter with multiple extensions anyway, and I just changed that. This also combines the extension string with a MIME type (currently unused but might be needed in the future), JavaFX DataFormat (used for clipboard access in copy/paste), Java class (also used for clipboard access) and description (for tooltips later). All of this is to set up a feature I have in mind...

Signed-off-by: Emi <dealuminis@gmail.com>