Normalize dialog controller specification and switch to FxUtils.FXML
A few old dialogs were breaking due to a change in FxUtils.FXML (I think...? Or maybe just my JDK...?), so I went through and normalized things, mostly. The ImageExportUI isn't a proper JavaFX thing yet, and anyway I want it to be a private class which JavaFX doesn't like.

Signed-off-by: Emi <dealuminis@gmail.com>