Fix concurrent modification exception with tags when loading the collection
This was bugging me for the longest time -- sometimes when loading the deckbuilder, the collection would just... not load. It took me a while to figure out why -- it turns out, tasks submitted to the common pool can fail silently, including the collection load task. In this particular case, it was failing because loading the collection was accessing and iterating the global tags collections, while the tags collections was being loaded, resulting in a concurrent modification exception on the thread setting up the collection.

I unthreaded the tag loading and synchronized the collection loading on the tag map, to be doubly sure. Also added better logging of this failure. Still need to go through all uses of ForkJoinPool.commonPool to make sure EVERYTHING is logged...

Signed-off-by: Emi <dealuminis@gmail.com>