FilteredGroupedModel refactor \o/
It's finally here! God, cleaning up these code changes has taken me way too long. Okay, so, summary: I wrote a class called FilteredGroupedModel, which takes any ObservableList of elements T as a source, and a function which extracts a set of groups G from each element as well as a predicate on each element T. Whenever the source list changes, the FGM filters and groups the source elements in parallel (if there are enough of them), yielding a map of G -> List<T>.

All this has rather dramatically simplified(?) CardView, which no longer needs to perform grouping itself, and massively improves the performance of searches on the collection on multicore processors. My computer saw a reduction from about 10s to group the entire collection the first time, down to about 2s!

There's one glaring flaw, however: Changes to the source list always provoke a full regroup of the model, no matter how minute. In a future update, I hope to revise this; see the self-discussion in `FilteredGroupedModel.sourceChanged` for some good motivation! Still, I'm very happy with this improvement.

Signed-off-by: Emi <dealuminis@gmail.com>