Add new updater dialog and Updateable interface to be implemented by various data sources
As a part of adding support for tags from Scryfall, I realized things other than just the data source might want or need to update local caches. Which led to this: a nice grid display of things requiring updates, buttons to update them individually, and a button to update them all.

Signed-off-by: Emi <dealuminis@gmail.com>