Collection tags refactor: Multiplex tag providers and generally rebuild tags system
The old tags system was pretty quick to whip up, but lacked in any kind of reasonable extensibility. A number of my users requested access to Scryfall's tags, and while the precise approach to accomplishing that is still up for debate, I still needed to rework the tagging system if I wanted to do that. So I have! This commit does introduce a fair few TODOs (specifically around saving/loading tags without progress indication, mainly -- although so far that process is fast enough that it doesn't matter), but I'm happy with the result as apparent to users, so committing it.

Signed-off-by: Emi <dealuminis@gmail.com>