Fix cardview scroll reset and coalesce relayouts on group model changes
Not sure exactly when this happened, I'll have to find the commit. But changes were causing the scroll of cardpanes to reset, and it was bugging me; this was why.

This change also accomplishes something else that was bugging me: multiple rapid changes to the grouped model were triggering layout several times in a row. This was wholly unnecessary (and honestly still is), but a consequence of how JavaFX wrote their ObservableMap change events. By implementing a self-canceling generation system as with rendering, we can cancel the layout process if a second layout is scheduled in quick succession.

Signed-off-by: Emi <dealuminis@gmail.com>