Replace "Collapse Duplicates" and "Show Versions Separately" with Uniqueness and deduplication in FilteredGroupedModel
I've hated how those old settings interacted, and especially Show Versions Separately was implemented in a very cheap and bad way, which caused a lot of problems with print-specific filters (e.g. set, rarity). It also always seemed to me that those two settings were different ways of expressing the same thing -- lumping together prints -- and it would make more sense to do them at the same time. This approach gives the added boon of doing the "preferred print" logic and the deduplication in a multithreaded environment, which should noticeably improve performance.

I've also added the "unique across groups" setting, which alters the deduplication behavior to be global vs. per-group. This is actually less efficient (and the implementation is jank), but I usually don't like to see more than one of a card, especially with rarity grouping on the collection, because rarity-boosted prints tend to be promotional or otherwise fancy border. Default behavior is to dedupe globally (even if it is less efficient), but there's a toggle for it in the Unique menu.

Downside, I'm gonna have to rename FilteredGroupedModel to FilteredDeduplicatedGroupedModel. Or SwissArmyModel. Or MultitoolModel. Or ListDemuxer. I haven't settled on what I'm going to call it, but it's become very complicated.

Signed-off-by: Emi <dealuminis@gmail.com>